model1/requirements.txt
accelerate
requests
torch
torchvision
transformers
pillow
scipy
segment-anything
opencv-python
flask